Rename enemy module to engine
